adicionado tags e select2

+++ projeto/projetos/views.py
@@ -1,4 +1,6 @@
 from django import shortcuts
+from django.views import generic
+from django.contrib import messages
 
 from . import forms, models
 
@@ -18,6 +20,7 @@ def criar(request):
         projeto = form.save()
         inline_form.instance = projeto
         inline_form.save()
+        messages.success(request, 'Projeto criado com sucesso')
         return shortcuts.redirect('/')
 
     return shortcuts.render(request, 'projetos.html', {
@@ -42,6 +45,7 @@ def editar(request, pk=None):
     if form.is_valid() and inline_form.is_valid():
         form.save()
         inline_form.save()
+        messages.info(request, 'Projeto Atualizado com sucesso')
         return shortcuts.redirect('/')
 
     return shortcuts.render(request, 'projetos.html', {
@@ -52,5 +56,24 @@ def editar(request, pk=None):
 
 def remover(request, pk=None):
     projeto = shortcuts.get_object_or_404(models.Projeto, pk=pk)
+    if request.method == 'GET':
+        return shortcuts.render(request, '', {
+            'projeto': projeto
+        })
+
     projeto.delete()
     return shortcuts.redirect('/')
+
+
+class RemoverProjetoView(generic.DeleteView):
+
+    template_name = 'remover_projeto.html'
+    model = models.Projeto
+    success_url = '/'
+    
+    def delete(self, request, *args, **kwargs):
+        messages.warning(request, 'Projeto removido com sucesso')
+        return super().delete(request, *args, **kwargs)
+
+
+remover = RemoverProjetoView.as_view()
